backend/python-backend/requirements.txt
fastapi
uvicorn
pika
python-dotenv
requests
